requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.3
pytest==7.2.2
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-ordering==0.6
pytest-parallel==0.1.1
pytest-rerunfailures==11.1.2
requests==2.28.2
tblib==1.7.0
tomli==2.0.1
urllib3==1.26.15
xlrd==1.2.0
